skills/trends-scraper/requirements.txt
apify-client>=2.0.0
openai>=2.0.0
psycopg2-binary>=2.9
python-dotenv>=1.0.0
pytrends>=4.9.2
requests>=2.31
